cassandra: add label - intent for decommission
We add a new label to our constants, called
`cassandra.rook.io/decommissioned`. When it is present on a
member's ClusterIP Service with the value "false", it expresses
the intent of the controller to decommission the member. When it is
present with the value "true", it means that decommission has completed.

Signed-off-by: Yannis Zarkadas <yanniszarkadas@gmail.com>